backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
werkzeug==3.0.1
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
pytest==7.4.4
pytest-flask==1.3.0
anthropic>=0.40.0
